% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcNormS}
\alias{calcNormS}
\title{Calculate normalised S matrix}
\usage{
calcNormS(D, ncols)
}
\arguments{
\item{D}{list of data matrices}

\item{ncols}{number of columns of input matrices}
}
\description{
Calculate normalised S matrix
}
