% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hogsvdR.R
\name{hogsvd.rArmadillo}
\alias{hogsvd.rArmadillo}
\title{Compute the HO GSVD decomposition of a list of matrices using c acceleration}
\usage{
hogsvd.rArmadillo(D)
}
\arguments{
\item{D}{a list of matrices to compute the GSVD decomposition on}
}
\value{
A list of U, Sigma, V,  Lambda and S. U and Sigma are lists of matrices
}
\description{
Compute the HO GSVD decomposition of a list of matrices using c acceleration
}
