% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hogsvdR.R
\name{hogsvd}
\alias{hogsvd}
\title{Compute the HO GSVD decomposition of a list of matrices}
\usage{
hogsvd(D, method = "rsimple")
}
\arguments{
\item{D}{a list of matrices to compute the GSVD decomposition on}

\item{method}{specification of internal function to use to compute HOGSVD, currently only 'rsimple' supported}
}
\value{
A list of U, Sigma and V. U and Sigma are lists of matrices
}
\description{
Compute the HO GSVD decomposition of a list of matrices
}
\examples{
N <- 3
nrow <- c(10,10,10)
ncol <- 10
s <- 1:N
D <- lapply(s, function(x) {matrix(rnorm(n=nrow[x]*ncol,mean = 0, sd =10),nrow[x],ncol)})
res <- hogsvd(D)

D.reconstruct <- lapply(1:N, function(n) { res$U[[n]] \%*\% diag(res$Sigma[[n]]) \%*\% t(res$V) })
}
