% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hogsvdR.R
\name{hogsvd.rArmadillo}
\alias{hogsvd.rArmadillo}
\title{Compute the Higher-order generalised singular value decomposition (HOGSVD) of a list of matrices with RcppArmadillo}
\usage{
hogsvd.rArmadillo(D)
}
\arguments{
\item{D}{a list of matrices to compute the GSVD decomposition on}
}
\value{
A list of U, Sigma, V,  Lambda and S. U and Sigma are lists of matrices
}
\description{
Compute the Higher-order generalised singular value decomposition (HOGSVD) of a list of matrices with RcppArmadillo
}
\examples{
N <- 3
nrow <- c(10,10,10)
ncol <- 10
s <- 1:N
D <- lapply(s, function(x) {matrix(rnorm(n=nrow[x]*ncol,mean = 0, sd =10),nrow[x],ncol)})

# Perform HO GSVD on the example
res <- hogsvd.rArmadillo(D)
}
